Separate Jimple Pretty Printer from Code Generator
